Separate runner and index files
